requirements.txt
streamlit==1.31.0
pandas==2.1.4
openpyxl==3.1.2
pillow==10.2.0
xlsxwriter==3.1.9
python-dateutil==2.8.2
numpy==1.24.3
gspread==5.12.0
google-auth==2.23.4